refactor: RandomNumber 클래스 model 패키지로 이동
